restructure backend and update naming convention
