--- backend/src/config/firebase-admin.ts
@@ -1,7 +1,0 @@
-var admin = require('firebase-admin');
-
-var serviceAccount = require('serviceAccountKey.json');
-
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-});
